+++ src/main/java/ru/yandex/practicum/filmorate/util/IdGen.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 public class IdGen {
     @Getter
     @Setter
-    private int id = 0;
-    public int nextId() {
+    private long id = 0;
+    public long nextId() {
         return ++id;
     }
 }
